backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiomysql==0.2.0
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
python-socketio==5.11.4
pydantic-settings==2.5.2
cryptography==43.0.1
aiofiles==24.1.0
